fix-checkbox-editbox-conflict
